Move types.tsx to util
